pulled out the image metadata into it's own div for future use
