+++ src/upload-page/upload-page-dropzone-img-thumbnail.js
